backend/requirements.txt
fastapi[standard]
uvicorn
python-dotenv 
pydantic
mysql-connector-python
openai
elevenlabs
os
